Initial working WindowsProcess

+++ windowslib/src/main/resources/scala-native/windows/console.c
@@ -1,11 +1,11 @@
 #include "windows.h"
 
-unsigned long scalanative_win32_console_std_in_handle() {
+DWORD scalanative_win32_console_std_in_handle() {
     return STD_INPUT_HANDLE;
 }
-unsigned long scalanative_win32_console_std_out_handle() {
+DWORD scalanative_win32_console_std_out_handle() {
     return STD_OUTPUT_HANDLE;
 }
-unsigned long scalanative_win32_console_std_err_handle() {
+DWORD scalanative_win32_console_std_err_handle() {
     return STD_ERROR_HANDLE;
 }
